laravel 6.0

+++ database/migrations/2015_11_14_161541_create_finance_records_table.php
@@ -13,13 +13,13 @@ class CreateFinanceRecordsTable extends Migration
     public function up()
     {
         Schema::create('finance_records', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->date('date');
             $table->string('title',255);
             $table->decimal('amount', 8, 2);
 //            $table->morphs('another')->nullable(); // source or destination
             $table->string('another_type')->nullable();
-            $table->integer('another_id')->unsigned()->nullable();
+            $table->bigInteger('another_id')->unsigned()->nullable();
             $table->string('memo',5000);
             $table->timestamps();
         });
