+++ database/migrations/2016_02_28_190148_create_supplements_table.php
@@ -15,8 +15,8 @@ class CreateSupplementsTable extends Migration
         // 附加内容，不同于存为数据表的“摘抄”或“推荐文章”，这里的格式是自由格式，方便编辑发挥
         //
         Schema::create('supplements', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('article_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('article_id')->unsigned();
             $table->string('body',5000);
             $table->text('md');
             $table->string('codes',500)->nullable();
