+++ database/migrations/2016_07_23_000602_create_books_table.php
@@ -13,9 +13,9 @@ class CreateBooksTable extends Migration
     public function up()
     {
         Schema::create('books', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-            $table->integer('volume_id')->unsigned()->nullable(); // 允许非vol的视频
+            $table->bigInteger('volume_id')->unsigned()->nullable(); // 允许非vol的视频
             $table->enum('type',[
                 'first',
                 'normal',
@@ -23,10 +23,10 @@ class CreateBooksTable extends Migration
 
             $table->boolean('version');
 //            $table->integer('volumn_id')->unsigned()->nullable();
-            $table->integer('order')->unsigned()->nullable(); // for version
-            $table->integer('book_id')->unsigned()->nullable(); // for version
+            $table->bigInteger('order')->unsigned()->nullable(); // for version
+            $table->bigInteger('book_id')->unsigned()->nullable(); // for version
 
-            $table->integer('image_id')->unsigned()->nullable();
+            $table->bigInteger('image_id')->unsigned()->nullable();
             $table->string('slug',100)->unique()->nullable();
             $table->string('name',50);
             $table->string('other_names',300)->nullable(); // 其它汉语名字,格式为: 钢锯岭\钢铁英雄(台)
@@ -35,15 +35,15 @@ class CreateBooksTable extends Migration
             $table->string('names',300)->nullable(); // 其它语言的名字,格式为: ja\\日语1\日语2;fr\\法语1
 
             $table->string('author',150)->nullable();
-            $table->integer('author_id')->unsigned()->nullable();
+            $table->bigInteger('author_id')->unsigned()->nullable();
 
             $table->string('time')->nullable();
             $table->date('date')->nullable();
 
             $table->string('intro',3000)->nullable(); // markdown
             $table->string('desc',200)->nullable();
-            $table->integer('tip_id')->unsigned()->nullable(); // for version
-            $table->integer('errata_id')->unsigned()->nullable(); // for version
+            $table->bigInteger('tip_id')->unsigned()->nullable(); // for version
+            $table->bigInteger('errata_id')->unsigned()->nullable(); // for version
 
             $table->enum('integrity',['-','?','✔','✖'])->nullable();
