+++ database/migrations/2016_10_07_102009_create_volumes_table.php
@@ -14,13 +14,13 @@ class CreateVolumesTable extends Migration
     public function up()
     {
         Schema::create('volumes', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('title',50);
 
-            $table->integer('column_id')->unsigned();
-            $table->integer('no')->unsigned();
+            $table->bigInteger('column_id')->unsigned();
+            $table->bigInteger('no')->unsigned();
 
-            $table->integer('person_id')->unsigned()->nullable();
+            $table->bigInteger('person_id')->unsigned()->nullable();
             $table->timestamps();
 
             $table->index(['column_id','no'])->unique();
