+++ database/migrations/2016_12_02_115508_create_videos_table.php
@@ -14,15 +14,15 @@ class CreateVideosTable extends Migration
     public function up()
     {
         Schema::create('videos', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-            $table->integer('volume_id')->unsigned()->nullable(); // 允许非vol的视频
+            $table->bigInteger('volume_id')->unsigned()->nullable(); // 允许非vol的视频
             $table->enum('type',[
                 'first',
                 'normal',
                 ])->default('first');
 
-            $table->integer('image_id')->unsigned()->nullable();
+            $table->bigInteger('image_id')->unsigned()->nullable();
             $table->string('slug',100)->unique()->nullable();
             $table->string('name',50);
             $table->string('other_names',300)->nullable(); // 其它汉语名字,格式为: 钢锯岭\钢铁英雄(台)
@@ -33,16 +33,16 @@ class CreateVideosTable extends Migration
 //            $table->string('pic',150);
 
             $table->string('author',50);
-            $table->integer('author_id')->unsigned()->nullable();
+            $table->bigInteger('author_id')->unsigned()->nullable();
 
             $table->string('time')->nullable();
             $table->date('date');
 
             $table->string('intro',300);
             $table->string('desc',200)->nullable();
-            $table->integer('tip_id')->unsigned()->nullable(); // for version
+            $table->bigInteger('tip_id')->unsigned()->nullable(); // for version
             // 背景资料
-            $table->integer('behind_id')->unsigned()->nullable(); // for version
+            $table->bigInteger('behind_id')->unsigned()->nullable(); // for version
 
             $table->timestamps();
         });
