+++ database/migrations/2018_07_04_010026_create_trees_table.php
@@ -14,7 +14,7 @@ class CreateTreesTable extends Migration
     public function up()
     {
         Schema::create('trees', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('slug', 100)->unique()->nullable();
             $table->string('name', 50);
@@ -27,9 +27,9 @@ class CreateTreesTable extends Migration
             $table->string('contact', 1000)->nullable(); // markdown
             $table->string('buy', 1000)->nullable(); // markdown
 
-            $table->integer('editor_id')->unsigned()->default(1); // 编辑是谁
+            $table->bigInteger('editor_id')->unsigned()->default(1); // 编辑是谁
             $table->integer('menu_item_id')->unsigned();
-            $table->integer('image_id')->unsigned()->nullable();
+            $table->bigInteger('image_id')->unsigned()->nullable();
 
             $table->timestamps();
         });
